feat(variables): add EditVariableModal component for variable editing functionality
