+++ frontend/src/library.js
@@ -82,6 +82,7 @@ import FilterNavBar from '@/components/task/filter/FilterNavBar.vue'
 import FilterNavCollapsed from '@/components/task/filter/FilterNavCollapsed.vue'
 import ProcessView from '@/components/process/ProcessView.vue'
 import AddVariableModal from '@/components/process/modals/AddVariableModal.vue'
+import EditVariableModal from '@/components/process/modals/EditVariableModal.vue'
 import DeleteVariableModal from '@/components/process/modals/DeleteVariableModal.vue'
 import BpmnViewer from '@/components/process/BpmnViewer.vue'
 import InstancesTable from '@/components/process/tables/InstancesTable.vue'
@@ -192,6 +193,7 @@ const registerComponents = function(app) {
   app.component('filter-nav-collapsed', FilterNavCollapsed)
   app.component('process-view', ProcessView)
   app.component('add-variable-modal', AddVariableModal)
+  app.component('edit-variable-modal', EditVariableModal)
   app.component('delete-variable-modal', DeleteVariableModal)
   app.component('bpmn-viewer', BpmnViewer)
   app.component('instances-table', InstancesTable)
@@ -307,6 +309,7 @@ export {
   FilterNavCollapsed,
   ProcessView,
   AddVariableModal,
+  EditVariableModal,
   DeleteVariableModal,
   BpmnViewer,
   InstancesTable,
@@ -396,6 +399,6 @@ export {
   fetchDecisionsIfEmpty,
   setupTaskNotifications,
   parseXMLDocumentation,
-  
+
   initEmbeddedForm
 }
